requirements.txt
langchain==0.2.10
langchain-core==0.2.22
langchain-google-genai==0.0.9
google-generativeai==0.5.4
python-dotenv==1.0.1
fastapi==0.111.1
langserve==0.2.2
sse_starlette==2.1.2
httpx==0.24.1